refactor: split core element.ts (interpolation engine + transform application into sibling modules) and polyline.ts (curve renderers module) — pure moves, exports unchanged
